Still getting stuff up to a working level
Added a new connections.py file, which will be handling all of
the actual connection logic in the future. Will ensure safe sending
and receipt of complete messages.
Added a stream handler for logging to output important log messages
to the stdout stream in addition to logging them in a file.
Renamed scanning.py to scraping.py to not confuse myself when
I get around to adding proper scanning/parsing for IRC messages,
as well as to more properly align with what it does.