Finishing up connection functionality
Added reconnect code to connection.py, will be called
if a connection attempt times out, and will reattempt
until the program is halted or a connection is made.

Added receiving logic to connection.py, should be properly
caching data that's missing the CRLF terminator, and not
returning a non-empty string until the full message has
been received from the socket.