Fixed receive logic, added some barebones functionality to the bot
Receive caching has been moved to a class attribute because the way
I was trying to do it I would have had to clear the default list
argument iteratively and then append to it for the cache to persist
between function calls. Using a class attrib is more efficient
and generally more attractive than getting it to work in function.

Added barebones functionality to the bot itself, none of the new
methods are in their final state. Join/Quit/Send are obviously
missing more logic to make them more compliant to the standard.
Connect is more spec compliant than my last iteration, but
checking responses is going to remain very hacky until I have
written up a proper scanner/parser for IRC messages.