Improved log file generation and error checking logic
While doing a quick check to see if the code would still run
before working on it, I hit an error I didn't expect.
Turns out, if the logs directory is missing, as it is absent
in the repository, the program would immediately fail out
because it can't find the log file. After learning the issue
was that the directory was missing I did some more testing
to think of ways to break the logging initialization.

The results are as follows:
If the config directory is missing, it will throw a KeyError
which would not have been handled by the old checks, this is
now handled.

If there was no logging.ini at all inside the config directory,
the software would fail out on attempting to rename the old
config. This is also now handled.

Per above, if the logs directory is missing it is now created,
and the logging library will create the log file itself if
it is missing. It may be possible that it will throw an error
if the file is in read only or inaccessible entirely, but
for now the adjusted checks should handle issues fine