Adjusted logging functionality and initialization
Logging now additionally logs messages of "WARNING" level and upwards
in a separate errors file by default

The main file was restructured to cut down on each function's complexity
Functions were also renamed to better indicate their usage.

If the logfile exists but is read-only the system will fail out with no
recovery. I have yet to determine an appropriate way to handle this
situation so this will remain for the time being.